# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kinora API",
    "summary": "Generation-on-scroll showrunner backend.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me/prefs": {
      "get": {
        "tags": [
          "prefs"
        ],
        "summary": "My Directing Style",
        "description": "The reader's accumulated directing style across all their books (\u00a78.6).",
        "operationId": "my_directing_style_api_me_prefs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectingStyleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "prefs"
        ],
        "summary": "Reset My Directing Style",
        "description": "Clear the reader's learned directing style everywhere (global reset, \u00a78.6).",
        "operationId": "reset_my_directing_style_api_me_prefs_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPrefsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/books/{book_id}/prefs": {
      "get": {
        "tags": [
          "prefs"
        ],
        "summary": "Book Directing Style",
        "description": "The directing style learned for one book \u2014 what its shots default to (\u00a78.6).",
        "operationId": "book_directing_style_api_books__book_id__prefs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectingStyleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prefs"
        ],
        "summary": "Reset Book Directing Style",
        "description": "Clear the directing style learned for one book (\u00a78.6).",
        "operationId": "reset_book_directing_style_api_books__book_id__prefs_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPrefsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectingPriorView": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "bias": {
            "type": "number",
            "title": "Bias"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied"
          },
          "applied_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Value"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "bias",
          "weight",
          "label",
          "detail",
          "applied"
        ],
        "title": "DirectingPriorView",
        "description": "One learned directing prior, projected as plain language for the panel.\n\n``bias`` is the signed strength (negative = slower/cooler/closer, positive =\nfaster/warmer/wider); ``applied`` is whether it is strong enough to default a\nfuture shot; ``applied_value`` is the concrete default it sets (``slow`` /\n``wide`` / \u2026) when applied."
      },
      "DirectingStyleResponse": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "priors": {
            "items": {
              "$ref": "#/components/schemas/DirectingPriorView"
            },
            "type": "array",
            "title": "Priors"
          }
        },
        "type": "object",
        "required": [
          "scope"
        ],
        "title": "DirectingStyleResponse",
        "description": "The reader's accumulated directing style for a scope (\u00a78.6).\n\n``scope`` is ``\"book\"`` (this title) or ``\"user\"`` (all books). Empty\n``priors`` means nothing has been learned yet \u2014 the panel shows its zero-state."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResetPrefsResponse": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "cleared": {
            "type": "integer",
            "title": "Cleared",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "scope"
        ],
        "title": "ResetPrefsResponse",
        "description": "Acknowledgement that learned priors were cleared (\u00a78.6)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}